Add run_all script and rename files to sequence
