+++ pyproject.toml
@@ -87,13 +87,11 @@ select = [
     "E",    # pycodestyle
     "F",    # pyflakes
     "ISC",  # implicit string concatenation
-#    "I",    # sort imports
+#    "I",    # sort imports, commented out because sorted by isort.
     "UP",   # alert you when better syntax is available in your python version
     "RUF",  # the ruff developer's own rules
 ]
 ignore = [
-    "ANN101", # Supress missing-type-self.
-    "ANN102", # Supress missing-type-cls.
     "ANN202", # Don't requiere return type annotation for private functions.
     "ANN401", # Allow type annotation with type Any.
     "D100",   # Supress undocumented-public-module. Only doc of public api required.
