+++ tests/test_kildomat.py
@@ -3,15 +3,17 @@ from pathlib import Path
 import pandas as pd
 from pytest_mock import MockerFixture
 
-from src.notebooks.kildomat import process_observations
-from src.notebooks.kildomat import process_weather_stations
+from src.notebooks.b_kildomat import process_observations
+from src.notebooks.b_kildomat import process_weather_stations
 
 
 def test_process_weather_stations(mocker: MockerFixture) -> None:
     source_file = Path(__file__).parent / "testdata" / "weather_stations_v1.json"
 
     # Mock the function write_parquet_file. Then only a dummy target_dir is needed.
-    mock_write_parquet_file = mocker.patch("src.notebooks.kildomat.write_parquet_file")
+    mock_write_parquet_file = mocker.patch(
+        "src.notebooks.b_kildomat.write_parquet_file"
+    )
     dummy_target_dir = Path(__file__).parent
 
     process_weather_stations(source_file, target_dir=dummy_target_dir)
@@ -32,7 +34,9 @@ def test_process_observations(mocker: MockerFixture) -> None:
     source_file = Path(__file__).parent / "testdata" / "observations_p2011_v1.json"
 
     # Mock the function write_parquet_file. Then only a dummy target_dir is needed.
-    mock_write_parquet_file = mocker.patch("src.notebooks.kildomat.write_parquet_file")
+    mock_write_parquet_file = mocker.patch(
+        "src.notebooks.b_kildomat.write_parquet_file"
+    )
     dummy_target_dir = Path(__file__).parent
 
     process_observations(source_file, target_dir=dummy_target_dir)
