+++ tests/test_versions_get_latest_bucket.py
@@ -7,8 +7,9 @@ from functions.ssbplatforms import is_dapla
 from functions.versions import get_latest_file_version
 
 
-root_dir = settings.product_root_dir.removesuffix(f"/{settings.short_name}")
-PREFIX = f"{root_dir}/temp/testcase/versiontest"
+if is_dapla():
+    root_dir = settings.product_root_dir.removesuffix(f"/{settings.short_name}")
+    PREFIX = f"{root_dir}/temp/testcase/versiontest"
 
 
 @pytest.mark.skipif(
